ex 1

+++ singly_linked_lists/1-list_len.c
@@ -10,7 +10,7 @@
  */
 size_t list_len(const list_t *h)
 {
-	unsigned int elements = 0;
+	size_t elements = 0;
 
 	while (h)
 	{
